services/requirements-common.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
psycopg[binary]==3.2.9
google-cloud-storage==2.19.0
google-cloud-pubsub==2.29.0
python-multipart==0.0.20
pypdf==5.8.0
Pillow==11.3.0
pytesseract==0.3.13
